requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==1.6.0
cycler==0.11.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.19.0
idna==3.3
importlib-metadata==4.11.1
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stable-baselines3==1.4.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.2
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
